patch for platform.c
Below is the lastest patch for src/tcsd/platform.c.  This fixes a few problems with
accessing the utmp database.  I still think the Linux code is horribly broken (the loop
logic doesn't seem to have an exit and getutid is not an iterating function), but
I don't have a way to verify it right now so I trust someone else will take a closer look.

I did this patch against the latest code in the git tree, not the 1.3.1. tarball source.

Signed-off-by: Wyllys Ingersoll <wyllys.ingersoll@sun.com>