pthread_t portability fix
This patch minimizes the effort to port TrouSerS on Windows
(please refer to previous mail sent by Giovanni Cabiddu).
The problem is that while in Linux the type "pthread_t"
is defined as "unsigned long int", in Windows it
corresponds to a structure. As consequence it is not
possible for example, when compiling the code for Windows,
to compare the field "thread_id" of " tcsd_thread_data"
with THREAD_NULL. (which is equal to 0).

In this patch i modified the definition of the
tcsd_thread_data structure, by making the field
"thread_id" as pointer and i inserted the allocation/deallocation
of the data (the proper size is given by THREAD_TYPE).
In this way, finding an available slot in the array of
"tcsd_thread_data" structure will consist in verifying
that "thread_id" is NULL.

Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>
Signed-off-by: Rajiv Andrade <srajiv@linux.vnet.ibm.com>