TCSD error handling improvements
This patch does a better cleanup if an error happens in the
function "tcsd_thread_create". In particular the "hostname"
string may not be freed and the socket is not closed if the
functions THREAD_ATTR_INIT, THREAD_ATTR_SETJOINABLE or THREAD_CREATE
return an error.
Further, "data->hostname" in the function "tcsd_thread_run" is freed
even if "TCSD_SINGLE_THREAD_DEBUG" is defined since the pointer
assignment, in the function "tcsd_thread_create", is made in either
cases.

Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>
Signed-off-by: Rajiv Andrade <srajiv@linux.vnet.ibm.com>