support for -c <configfile> command line option
This patch adds support for a -c|--config <configfile> command line
option that allows one to start the tcsd with a configuration file other
than the default configuration file.

- TCSD_CONFIG_FILE was renamed to TCSD_DEFAULT_CONFIG_FILE
- tcsd_config_file now holds the filename of the config file
- all occurrences of TCSD_CONFIG_FILE were replaced with
tcsd_config_file
- a '-c|--config <configfile>' command line option was introduced and
the optarg is stored in tcsd_config_file; if no such option is use,
TCSD_DEFAULT_CONFIG_FILE is stored into tcsd_config_file

There was a problem with the handling of the 'h' option which I fixed
also -- it now again displays the usage.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>
Signed-off-by: Rajiv Andrade <srajiv@linux.vnet.ibm.com>