SF #3103977 bug fix
The Trousers blob handling (trspi/trousers.c) seems to not allocate memory for a PCRInfo member in Trspi_Unload_CERTIFY_INFO. This can cause a program segfault if this function is used to read in a valid TPM_CERTIFY_INFO structure from a blob directly. Attached patch changes allocation for this field. I will also note that there seem to be other places where a PCRInfo field is loaded with Trspi_UnloadBlob. This seems wrong, given that the result is then not a valid TPM_CERTIFY_INFO pointer (variable length fields will not line up as one traverses the pointer tree).

Bug is not quite in the TSP Interface library, but this seemed the closest category.

[Replaced %u by %lu]
Signed-off-by: Rajiv Andrade <srajiv@linux.vnet.ibm.com>