trousers: fixed Tspi_Context_LoadKeyByUUID() function
This patch adds the necessary code in Tspi_Context_LoadKeyByUUID() for
determining the correct loadkey ordinal (it depends on the TPM version).
This information is used later in the function secret_PerformAuth_OIAP()
to set the authentication data in the TCS_LOADKEY_INFO structure, needed
to load the parent key.

Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>